feat(rec): Add theme selector for bat
With the environment variable "BAT_THEME" it is now possible
to change the default theme for bat from outside. This works
equivalent to the already existing environment variables
"BAT_STYLE" as well as "BAT_LANGUAGE" and has the same
effect as "BAT_STYLE" on the highlighting of content by the
shell function "syntax_highlighter()" as well as comments as
markdown by the shell function "markdown_highlighter()".

Signed-off-by: Stephan Linz <linz@li-pro.net>