feat(rec): allow shell alias during execution
In some cases it is desirable to manipulate the way as well as
the form of output of certain commands without unnecessarily
complicating the syntax of their invocation for recording
purposes. For this purpose, control by externally specified
aliases of the affected commands with extended command line
arguments hidden by the alias is a helpful method. However,
this method requires within the augmented script that the
expansion of given aliases is also active at any time of
a real evolution of a real command.

Signed-off-by: Stephan Linz <linz@li-pro.net>